Reorganize reconcile tests
