+++ internal/controllers/reconciliation/status_test.go
@@ -1,0 +1,196 @@
+package reconciliation
+
+import (
+	"testing"
+	"time"
+
+	apiv1 "github.com/Azure/eno/api/v1"
+	"github.com/Azure/eno/internal/controllers/aggregation"
+	testv1 "github.com/Azure/eno/internal/controllers/reconciliation/fixtures/v1"
+	"github.com/Azure/eno/internal/controllers/synthesis"
+	"github.com/Azure/eno/internal/reconstitution"
+	"github.com/Azure/eno/internal/testutil"
+	"github.com/google/uuid"
+	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/util/retry"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+// TestResourceReadiness proves that resources supporting readiness checks eventually have their state
+// mirrored into the resource slice.
+func TestResourceReadiness(t *testing.T) {
+	scheme := runtime.NewScheme()
+	corev1.SchemeBuilder.AddToScheme(scheme)
+	testv1.SchemeBuilder.AddToScheme(scheme)
+
+	ctx := testutil.NewContext(t)
+	mgr := testutil.NewManager(t)
+	upstream := mgr.GetClient()
+	downstream := mgr.DownstreamClient
+
+	// Register supporting controllers
+	rm, err := reconstitution.New(mgr.Manager, time.Millisecond)
+	require.NoError(t, err)
+	require.NoError(t, synthesis.NewRolloutController(mgr.Manager))
+	require.NoError(t, synthesis.NewStatusController(mgr.Manager))
+	require.NoError(t, synthesis.NewPodLifecycleController(mgr.Manager, defaultConf))
+	require.NoError(t, aggregation.NewStatusController(mgr.Manager))
+	require.NoError(t, synthesis.NewExecController(mgr.Manager, defaultConf, &testutil.ExecConn{Hook: func(s *apiv1.Synthesizer) []client.Object {
+		obj := &corev1.ConfigMap{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "test-obj",
+				Namespace: "default",
+				Annotations: map[string]string{
+					"eno.azure.io/readiness": "self.data.foo == 'baz'",
+				},
+			},
+			Data: map[string]string{"foo": s.Spec.Image},
+		}
+
+		gvks, _, err := scheme.ObjectKinds(obj)
+		require.NoError(t, err)
+		obj.GetObjectKind().SetGroupVersionKind(gvks[0])
+		return []client.Object{obj}
+	}}))
+
+	// Test subject
+	err = New(rm, mgr.DownstreamRestConfig, 5, testutil.AtLeastVersion(t, 15), time.Millisecond)
+	require.NoError(t, err)
+	mgr.Start(t)
+
+	// Any syn/comp will do since we faked out the synthesizer pod
+	syn := &apiv1.Synthesizer{}
+	syn.Name = "test-syn"
+	syn.Spec.Image = "bar"
+	syn.Spec.RolloutCooldown = &metav1.Duration{Duration: time.Millisecond}
+	require.NoError(t, upstream.Create(ctx, syn))
+
+	comp := &apiv1.Composition{}
+	comp.Name = "test-comp"
+	comp.Namespace = "default"
+	comp.Spec.Synthesizer.Name = syn.Name
+	require.NoError(t, upstream.Create(ctx, comp))
+
+	// Wait for resource to be created
+	obj := &corev1.ConfigMap{}
+	testutil.Eventually(t, func() bool {
+		obj.SetName("test-obj")
+		obj.SetNamespace("default")
+		err = downstream.Get(ctx, client.ObjectKeyFromObject(obj), obj)
+		return err == nil
+	})
+
+	// Initially readiness should be false
+	testutil.Eventually(t, func() bool {
+		slices, err := mgr.GetCurrentResourceSlices(ctx)
+		if err != nil {
+			t.Log(err)
+			return false
+		}
+		return len(slices[0].Status.Resources) > 0 && isNotReady(slices[0].Status.Resources[0])
+	})
+
+	testutil.Eventually(t, func() bool {
+		err = upstream.Get(ctx, client.ObjectKeyFromObject(comp), comp)
+		return err == nil && comp.Status.CurrentSynthesis != nil && comp.Status.CurrentSynthesis.Ready == nil
+	})
+
+	// Update resource to meet readiness criteria
+	err = retry.RetryOnConflict(testutil.Backoff, func() error {
+		err = upstream.Get(ctx, client.ObjectKeyFromObject(syn), syn)
+		syn.Spec.Image = "baz"
+		return upstream.Update(ctx, syn)
+	})
+	require.NoError(t, err)
+
+	// It should eventually be marked as ready
+	testutil.Eventually(t, func() bool {
+		slices, err := mgr.GetCurrentResourceSlices(ctx)
+		if err != nil {
+			t.Log(err)
+			return false
+		}
+		return len(slices[0].Status.Resources) > 0 && isReady(slices[0].Status.Resources[0])
+	})
+
+	// The composition should also be updated
+	testutil.Eventually(t, func() bool {
+		err = upstream.Get(ctx, client.ObjectKeyFromObject(comp), comp)
+		return err == nil && comp.Status.CurrentSynthesis != nil && comp.Status.CurrentSynthesis.Ready != nil
+	})
+
+	// Update resource to not meet readiness criteria
+	err = retry.RetryOnConflict(testutil.Backoff, func() error {
+		err = upstream.Get(ctx, client.ObjectKeyFromObject(syn), syn)
+		syn.Spec.Image = "bar"
+		return upstream.Update(ctx, syn)
+	})
+	require.NoError(t, err)
+
+	// The composition status should revert back to not ready when re-synthesized
+	testutil.Eventually(t, func() bool {
+		err = upstream.Get(ctx, client.ObjectKeyFromObject(comp), comp)
+		return err == nil && comp.Status.CurrentSynthesis != nil && comp.Status.CurrentSynthesis.Ready == nil
+	})
+}
+
+// TestReconcileStatus proves that reconciliation and deletion status are written to resource slices as expected.
+func TestReconcileStatus(t *testing.T) {
+	scheme := runtime.NewScheme()
+	corev1.SchemeBuilder.AddToScheme(scheme)
+	testv1.SchemeBuilder.AddToScheme(scheme)
+
+	ctx := testutil.NewContext(t)
+	mgr := testutil.NewManager(t)
+	upstream := mgr.GetClient()
+
+	rm, err := reconstitution.New(mgr.Manager, time.Millisecond)
+	require.NoError(t, err)
+
+	err = New(rm, mgr.DownstreamRestConfig, 5, testutil.AtLeastVersion(t, 15), time.Hour)
+	require.NoError(t, err)
+	mgr.Start(t)
+
+	comp := &apiv1.Composition{}
+	comp.Name = "test-comp"
+	comp.Namespace = "default"
+	require.NoError(t, upstream.Create(ctx, comp))
+	comp.ResourceVersion = "1"
+
+	slice := &apiv1.ResourceSlice{}
+	slice.Name = "test-slice"
+	slice.Namespace = comp.Namespace
+	require.NoError(t, controllerutil.SetControllerReference(comp, slice, upstream.Scheme()))
+	slice.Spec.Resources = []apiv1.Manifest{
+		{Manifest: `{ "kind": "ConfigMap", "apiVersion": "v1", "metadata": { "name": "test", "namespace": "default" } }`},
+		{Deleted: true, Manifest: `{ "kind": "ConfigMap", "apiVersion": "v1", "metadata": { "name": "test-deleted", "namespace": "default" } }`},
+	}
+	require.NoError(t, upstream.Create(ctx, slice))
+
+	now := metav1.Now()
+	err = retry.RetryOnConflict(testutil.Backoff, func() error {
+		upstream.Get(ctx, client.ObjectKeyFromObject(comp), comp)
+		comp.Status.CurrentSynthesis = &apiv1.Synthesis{
+			UUID:           uuid.NewString(),
+			Synthesized:    &now,
+			ResourceSlices: []*apiv1.ResourceSliceRef{{Name: slice.Name}},
+		}
+		return upstream.Status().Update(ctx, comp)
+	})
+	require.NoError(t, err)
+
+	// Status should eventually reflect the reconciliation state
+	testutil.Eventually(t, func() bool {
+		err = upstream.Get(ctx, client.ObjectKeyFromObject(slice), slice)
+		return err == nil && len(slice.Status.Resources) == 2 &&
+			slice.Status.Resources[0].Reconciled && !slice.Status.Resources[0].Deleted &&
+			slice.Status.Resources[1].Reconciled && slice.Status.Resources[1].Deleted
+	})
+}
+
+func isReady(state apiv1.ResourceState) bool    { return state.Ready != nil }
+func isNotReady(state apiv1.ResourceState) bool { return state.Ready == nil }
